% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\name{.vigiar_baixar_rj_year}
\alias{.vigiar_baixar_rj_year}
\title{Year-partitioned download for large tables}
\usage{
.vigiar_baixar_rj_year(tabela, ...)
}
\description{
Year-partitioned download for large tables
}
\keyword{internal}
