% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{.vigiar_converter_coluna}
\alias{.vigiar_converter_coluna}
\title{Convert a column to the appropriate R type}
\usage{
.vigiar_converter_coluna(x, type_code)
}
\arguments{
\item{x}{Column vector.}

\item{type_code}{Power BI data type code.}
}
\value{
Converted vector.
}
\description{
Convert a column to the appropriate R type
}
\keyword{internal}
