% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{.vigiar_detectar_outliers}
\alias{.vigiar_detectar_outliers}
\title{Detect outliers in numeric columns}
\usage{
.vigiar_detectar_outliers(dados, verbose = TRUE)
}
\description{
Detect outliers in numeric columns
}
\keyword{internal}
