% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{.vigiar_registrar_download}
\alias{.vigiar_registrar_download}
\title{Download history entry}
\usage{
.vigiar_registrar_download(tabela, n_rows, n_cols, elapsed, url)
}
\description{
Download history entry
}
\keyword{internal}
