% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{new_vigiar_diagnostic}
\alias{new_vigiar_diagnostic}
\title{Create a diagnostic object}
\usage{
new_vigiar_diagnostic(tabela, dados)
}
\arguments{
\item{tabela}{Table name}

\item{dados}{Input data frame}
}
\value{
A vigiar_diagnostic object
}
\description{
Create a diagnostic object
}
\keyword{internal}
