% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{new_vigiar_tbl}
\alias{new_vigiar_tbl}
\title{Create a typed VIGIAR tibble}
\usage{
new_vigiar_tbl(x, subclass = character(0), tabela = NULL, metadados = NULL)
}
\arguments{
\item{x}{A data frame or tibble.}

\item{subclass}{Character vector of additional class names.}

\item{tabela}{Original table name.}

\item{metadados}{List of metadata attributes.}
}
\value{
An object of class \code{vigiar_tbl} (and subclasses).
}
\description{
Create a typed VIGIAR tibble
}
\keyword{internal}
