% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{null-coalesce}
\alias{null-coalesce}
\alias{\%||\%}
\title{NULL-coalesce operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{A value.}

\item{y}{Fallback value if \code{x} is NULL.}
}
\value{
\code{x} unless \code{x} is NULL, then \code{y}.
}
\description{
Returns the left-hand value unless it is NULL, in which case
the right-hand value is returned.
}
