% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{print.vigiar_audit}
\alias{print.vigiar_audit}
\title{Print method for vigiar_audit}
\usage{
\method{print}{vigiar_audit}(x, ...)
}
\arguments{
\item{x}{A vigiar_audit object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for vigiar_audit
}
