% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{print.vigiar_audit_list}
\alias{print.vigiar_audit_list}
\title{Print method for vigiar_audit_list}
\usage{
\method{print}{vigiar_audit_list}(x, ...)
}
\arguments{
\item{x}{A vigiar_audit_list object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for vigiar_audit_list
}
