% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compliance.R
\name{print.vigiar_compliance}
\alias{print.vigiar_compliance}
\title{Print method for vigiar_compliance}
\usage{
\method{print}{vigiar_compliance}(x, ...)
}
\arguments{
\item{x}{A vigiar_compliance object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for vigiar_compliance
}
