% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{print.vigiar_diagnostic}
\alias{print.vigiar_diagnostic}
\title{Print a vigiar_diagnostic object}
\usage{
\method{print}{vigiar_diagnostic}(x, ...)
}
\arguments{
\item{x}{A vigiar_diagnostic object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print a vigiar_diagnostic object
}
