% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{print.vigiar_snapshot}
\alias{print.vigiar_snapshot}
\title{Print method for vigiar_snapshot}
\usage{
\method{print}{vigiar_snapshot}(x, ...)
}
\arguments{
\item{x}{A vigiar_snapshot object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Print method for vigiar_snapshot
}
