% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_exposicao_indoor}
\alias{process_exposicao_indoor}
\title{Process indoor exposure data}
\usage{
process_exposicao_indoor(dados, tipo = c("exposicao", "desfecho"), ...)
}
\arguments{
\item{dados}{Raw data frame from \code{vigiar_baixar("df_indoor")}
or \code{vigiar_baixar("df_indoor_desfecho")}.}

\item{tipo}{One of \code{"exposicao"} or \code{"desfecho"}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{vigiar_indoor} tibble.
}
\description{
Process indoor exposure data
}
