% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_fracao_atribuivel}
\alias{process_fracao_atribuivel}
\title{Process attributable fraction data}
\usage{
process_fracao_atribuivel(dados, ...)
}
\arguments{
\item{dados}{Raw data frame from \code{vigiar_baixar("tb_fracao")}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{vigiar_attributable_fraction} tibble.
}
\description{
Process attributable fraction data
}
