% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_ufs}
\alias{process_ufs}
\title{Process UF-level data}
\usage{
process_ufs(dados, contexto = "uf")
}
\arguments{
\item{dados}{Raw data frame.}

\item{contexto}{Description of the data context.}
}
\value{
A \code{vigiar_tbl}.
}
\description{
Generic processor for any UF-level VIGIAR data.
}
