% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{summary.vigiar_diagnostic}
\alias{summary.vigiar_diagnostic}
\title{Summary method for vigiar_diagnostic}
\usage{
\method{summary}{vigiar_diagnostic}(object, ...)
}
\arguments{
\item{object}{A vigiar_diagnostic object.}

\item{...}{Additional arguments (ignored).}
}
\description{
Summary method for vigiar_diagnostic
}
