% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{vigiar_baixar_principais}
\alias{vigiar_baixar_principais}
\title{Download main tables (convenience shortcut)}
\usage{
vigiar_baixar_principais()
}
\value{
Named list of tibbles.
}
\description{
Downloads 14 key tables covering all data categories.
}
