% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vigiar_cache_info}
\alias{vigiar_cache_info}
\title{List cached tables}
\usage{
vigiar_cache_info()
}
\value{
A tibble with cache contents: tabela, linhas, colunas, idade, checksum.
}
\description{
List cached tables
}
