% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vigiar_carregar_snapshot}
\alias{vigiar_carregar_snapshot}
\title{Load a snapshot from disk}
\usage{
vigiar_carregar_snapshot(caminho)
}
\arguments{
\item{caminho}{File path to an .rds snapshot.}
}
\value{
A \code{vigiar_snapshot} object.
}
\description{
Load a snapshot from disk
}
