% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditar.R
\name{vigiar_comparar_schema}
\alias{vigiar_comparar_schema}
\title{Compare live schema with dictionary}
\usage{
vigiar_comparar_schema()
}
\value{
A tibble comparing live vs documented columns per table.
}
\description{
Compare live schema with dictionary
}
