% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vigiar_convencoes}
\alias{vigiar_convencoes}
\title{Open VIGIAR conventions documentation}
\usage{
vigiar_convencoes()
}
\description{
Opens the online conventions page (equivalent to microdatasus
"Convencoes SIH-RD").
}
