% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vigiar_descrever_variavel}
\alias{vigiar_descrever_variavel}
\title{Describe a single VIGIAR variable}
\usage{
vigiar_descrever_variavel(dominio, variavel)
}
\arguments{
\item{dominio}{Data domain (see \code{vigiar_variaveis}).}

\item{variavel}{Standard variable name.}
}
\value{
Invisibly, the dictionary row for the variable.
}
\description{
Describe a single VIGIAR variable
}
