% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vigiar_esquema_carregar_lock}
\alias{vigiar_esquema_carregar_lock}
\title{Load a schema lock}
\usage{
vigiar_esquema_carregar_lock(caminho = "vigiar_schema_lock.json")
}
\arguments{
\item{caminho}{Path to a lock file.}
}
\value{
A \code{vigiar_schema_lock} object.
}
\description{
Load a schema lock
}
