% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log.R
\name{vigiar_historico_downloads}
\alias{vigiar_historico_downloads}
\title{Show download history}
\usage{
vigiar_historico_downloads()
}
\value{
A tibble with download history since package load.
}
\description{
Show download history
}
