% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vigiar_limpar_cache}
\alias{vigiar_limpar_cache}
\title{Clear the local cache}
\usage{
vigiar_limpar_cache(tabelas = NULL, max_age = NULL)
}
\arguments{
\item{tabelas}{Character vector of table names. \code{NULL} = all.}

\item{max_age}{Maximum age to keep (e.g., \code{86400} for 1 day).
\code{NULL} = no age filter.}
}
\description{
Clear the local cache
}
