% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostic.R
\name{vigiar_relatorio_diagnostico}
\alias{vigiar_relatorio_diagnostico}
\title{Generate a diagnostic report}
\usage{
vigiar_relatorio_diagnostico(diag)
}
\arguments{
\item{diag}{A vigiar_diagnostic object.}
}
\value{
Invisibly, the diagnostic object.
}
\description{
Prints a formatted diagnostic report to the console.
}
