% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumo.R
\name{vigiar_resumo}
\alias{vigiar_resumo}
\title{Summarise a VIGIAR tibble -- generic dispatcher}
\usage{
vigiar_resumo(x, ...)
}
\arguments{
\item{x}{A processed VIGIAR tibble.}

\item{...}{Additional arguments passed to specific summarisers.}
}
\value{
A tibble with summary statistics.
}
\description{
Summarise a VIGIAR tibble -- generic dispatcher
}
