% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resumo.R
\name{vigiar_resumo_pm25}
\alias{vigiar_resumo_pm25}
\alias{vigiar_resumo_saude}
\alias{vigiar_resumo_populacao}
\alias{vigiar_resumo_fracao_atribuivel}
\alias{vigiar_resumo_indoor}
\title{Summarise PM2.5 data}
\usage{
vigiar_resumo_pm25(x, ...)

vigiar_resumo_saude(x, ...)

vigiar_resumo_populacao(x, ...)

vigiar_resumo_fracao_atribuivel(x, ...)

vigiar_resumo_indoor(x, ...)
}
\arguments{
\item{x}{A \code{vigiar_pm25} tibble.}

\item{...}{Ignored.}
}
\value{
A tibble with descriptive statistics.
}
\description{
Summarise PM2.5 data
}
