% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\name{vigiar_rj_macrorregioes}
\alias{vigiar_rj_macrorregioes}
\title{List Rio de Janeiro health macro-regions}
\usage{
vigiar_rj_macrorregioes()
}
\value{
Character vector of the 9 macro-regions.
}
\description{
List Rio de Janeiro health macro-regions
}
