% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\name{vigiar_rj_regioes_saude}
\alias{vigiar_rj_regioes_saude}
\title{List Rio de Janeiro health regions}
\usage{
vigiar_rj_regioes_saude()
}
\value{
Character vector of health regions.
}
\description{
List Rio de Janeiro health regions
}
