% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cache.R
\name{vigiar_salvar_snapshot}
\alias{vigiar_salvar_snapshot}
\title{Save a snapshot to disk (RDS)}
\usage{
vigiar_salvar_snapshot(snapshot, caminho, overwrite = FALSE)
}
\arguments{
\item{snapshot}{A \code{vigiar_snapshot} object.}

\item{caminho}{File path (will add .rds if absent).}

\item{overwrite}{If \code{FALSE}, errors if file exists.}
}
\value{
Invisibly, the file path.
}
\description{
Save a snapshot to disk (RDS)
}
