% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vigiar_schema}
\alias{vigiar_schema}
\title{Show schema for a data domain}
\usage{
vigiar_schema(dominio = "all")
}
\arguments{
\item{dominio}{Data domain name.}
}
\value{
A tibble with schema information.
}
\description{
Returns a summary of all variables in a domain: names, types,
descriptions, and units.
}
