% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conexao.R
\name{vigiar_sessao_ativa}
\alias{vigiar_sessao_ativa}
\title{Check if a VIGIAR session is active}
\usage{
vigiar_sessao_ativa()
}
\value{
`TRUE` if a session exists, `FALSE` otherwise.
}
\description{
Check if a VIGIAR session is active
}
