\name{vigiar_tabelas}
\alias{vigiar_tabelas}
\title{Lista as tabelas disponíveis no dashboard VIGIAR}
\usage{
vigiar_tabelas()
}
\value{Vetor de caracteres com os nomes das tabelas.}
\description{
Retorna os nomes de todas as tabelas disponíveis para download.
}
\examples{
\dontrun{
vigiar_conectar()
vigiar_tabelas()
}
}
