% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditar.R
\name{vigiar_tabelas_documentadas}
\alias{vigiar_tabelas_documentadas}
\title{List all documented tables}
\usage{
vigiar_tabelas_documentadas()
}
\value{
Character vector of table IDs present in the dictionary.
}
\description{
List all documented tables
}
