% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary.R
\name{vigiar_variaveis}
\alias{vigiar_variaveis}
\title{List variables for a specific data domain}
\usage{
vigiar_variaveis(
  dominio = c("pm25", "populacao_exposta", "indicadores_saude", "fracao_atribuivel",
    "exposicao_indoor", "municipios", "all")
)
}
\arguments{
\item{dominio}{One of \code{"pm25"}, \code{"populacao_exposta"},
\code{"indicadores_saude"}, \code{"fracao_atribuivel"},
\code{"exposicao_indoor"}, \code{"municipios"}, or \code{"all"}.}
}
\value{
A tibble subset of the dictionary.
}
\description{
List variables for a specific data domain
}
