% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auditar.R
\name{vigiar_variaveis_orfas}
\alias{vigiar_variaveis_orfas}
\title{Find documented variables that don't exist in the live schema}
\usage{
vigiar_variaveis_orfas()
}
\value{
A tibble of orphan variables (empty if none).
}
\description{
Find documented variables that don't exist in the live schema
}
