\name{print.vigiar_compliance}
\alias{print.vigiar_compliance}
\title{Print a vigiar_compliance object}
\usage{
\method{print}{vigiar_compliance}(x, ...)
}
\arguments{
  \item{x}{A \code{vigiar_compliance} object.}
  \item{...}{Additional arguments (ignored).}
}
\description{
  Prints a compliance check report.
}
