% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_municipios}
\alias{process_municipios}
\title{Process municipality registry data}
\usage{
process_municipios(dados, ...)
}
\arguments{
\item{dados}{Raw data frame from \code{vigiar_baixar("df_muni")}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{vigiar_municipios} tibble.
}
\description{
Process municipality registry data
}
