% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process_populacao_exposta}
\alias{process_populacao_exposta}
\title{Process population exposure data}
\usage{
process_populacao_exposta(dados, ...)
}
\arguments{
\item{dados}{Raw data frame from \code{vigiar_baixar("pop")}.}

\item{...}{Additional arguments (ignored).}
}
\value{
A \code{vigiar_population} tibble.
}
\description{
Process population exposure data
}
