\name{vigiar-package}
\alias{vigiar-package}
\alias{vigiar}
\docType{package}
\title{Download Data from the VIGIAR Power BI Dashboard}
\description{
Provides functions to download air quality and health surveillance data
from Brazil's VIGIAR (Vigilancia em Saude Ambiental) Power BI dashboard.
}
\author{Maintainer: vigiar team <vigiar@example.com>}
\keyword{package}
