% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rj.R
\name{vigiar_rj_resumo}
\alias{vigiar_rj_resumo}
\title{Summarise Rio de Janeiro VIGIAR data}
\usage{
vigiar_rj_resumo(
  dados,
  agregacao = c("municipio", "macrorregiao", "regiao_saude")
)
}
\arguments{
\item{dados}{A processed VIGIAR tibble.}

\item{agregacao}{One of "municipio", "macrorregiao", or "regiao_saude".}
}
\value{
A tibble with summary statistics.
}
\description{
Summarise Rio de Janeiro VIGIAR data
}
